Add files via upload
The array_advance function checks if it is possible to advance from the start of the input array A to its last element, where the maximum number of steps that can be taken from a position is determined by the value of the element in the array at the current index. The function returns a boolean value indicating whether or not it is possible to do so.

Inside the while loop, we update the furthest_reached variable to be the maximum value between its current value and the sum of the value at the current index A[i] and the current index i. This calculates the furthest index that can be reached from the current index. We then increment i to move to the next index in the array.